refactor: Phase 13b–13e — decomposition of PresetView, lorebookState, ChatMessage, ChatInput
