+++ AGENTS.md
@@ -68,7 +68,7 @@ git checkout -b feat/memorybook-ui feat/multi-vector
 | Branch | Purpose | PR |
 |--------|---------|----|
 | `origin/dev` | Local mirror of upstream integration branch | No PR |
-| `feat/refactor-phase1-event-hub` | Phase 1–13a structural refactor (event hub, request ownership, composable extraction, transport split, use-case re-architecture, App.vue decomposition) | Not yet |
+| `feat/refactor-phase1-event-hub` | Phase 1–13e structural refactor (event hub, request ownership, composable extraction, transport split, use-case re-architecture, App.vue decomposition, PresetView.vue decomposition, lorebookState.js decomposition, ChatMessage.vue decomposition, ChatInput.vue decomposition) | Not yet |
 
 ### Historical (merged & deleted)
 - `feat/cloud-sync` → merged via PR #20
